+++ example/custom_backend/my_tavern_plugin/plugin.py
@@ -1,4 +1,3 @@
-import logging
 import pathlib
 from collections.abc import Iterable
 from os.path import abspath, dirname, join
@@ -15,6 +14,7 @@ from tavern.response import BaseResponse
 
 class Session:
     """No-op session, but must implement the context manager protocol"""
+
     def __enter__(self):
         pass
 
@@ -24,6 +24,7 @@ class Session:
 
 class Request(BaseRequest):
     """Touches a file when the 'request' is made"""
+
     def __init__(
         self, session: Any, rspec: dict, test_block_config: TestConfig
     ) -> None:
@@ -57,8 +58,6 @@ class Response(BaseResponse):
     ) -> None:
         super().__init__(name, expected, test_block_config)
 
-
-logger: logging.Logger = logging.getLogger(__name__)
 
 session_type = Session