+++ tavern/_core/exceptions.py
@@ -173,6 +173,12 @@ class ConcurrentError(TavernException):
     """Error while processing concurrent future"""
 
 
+class IncludedFileNotFoundError(TavernException, FileNotFoundError):
+    """A file referenced via file_body or include could not be found in the
+    include path (test file directory, current working directory, or
+    TAVERN_INCLUDE paths)"""
+
+
 class UnexpectedExceptionError(TavernException):
     """We expected a certain kind of exception in check_exception_raised but it was something
     else"""
@@ -184,3 +190,7 @@ class StarlarkError(TavernException):
 
 class DependencyMissingError(TavernException):
     """Tried to use some functionality for which the 'extra' had not been installed."""
+
+
+class TinctureError(TavernException):
+    """Badly specified tincture."""