+++ tavern/_plugins/grpc/__init__.py
@@ -1,18 +1,5 @@
 import warnings
 
-# Shut up warnings caused by proto libraries
-warnings.filterwarnings(
-    "ignore", category=DeprecationWarning, module="pkg_resources", lineno=2804
-)
-warnings.filterwarnings(
-    "ignore", category=DeprecationWarning, module="pkg_resources", lineno=2309
-)
-warnings.filterwarnings(
-    "ignore", category=DeprecationWarning, module="pkg_resources", lineno=2870
-)
-warnings.filterwarnings(
-    "ignore", category=DeprecationWarning, module="pkg_resources", lineno=2349
-)
-warnings.filterwarnings(
-    "ignore", category=DeprecationWarning, module="pkg_resources", lineno=20
-)
+# Suppress DeprecationWarnings from pkg_resources used by proto libraries.
+# These come from transitive dependencies and are not actionable here.
+warnings.filterwarnings("ignore", category=DeprecationWarning, module="pkg_resources")